#25766 Move the Quicktime generation settings into a hook Allows studios to easily override the default output settings. Also adds handling for the knob name change in Nuke 9.0 from `codec` to `meta_codec`
#29990 added setting to upload the movie to Shotgun
This allows reviewing the quick daily in the Shotgun web player. Note, this is false by default to maintain backwards compatibility.

#30791 new codec handling for Nuke 9 on linux
Nuke 9 removed ffmpeg from the `file_type` knob. It has been replaced with the mov64 writer which is now used and defaults to mjpeg codec to match the functionality on Mac and Windows (Note that this can now be modified with the new hook in this version).